# swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: urn:tos
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
host: localhost:8080
basePath: /
...
paths:
...
  /authors/findByName/{name}:
    get:
      tags:
        - author
      summary: findByName
      operationId: findByNameUsingGET
      produces:
        - '*/*'
      parameters:
        - name: name
          in: path
          description: name
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/AuthorDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
  /books/addBook:
    post:
      tags:
        - book
      summary: addBook
      operationId: addBookUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: book
          description: book
          required: true
          schema:
            $ref: '#/definitions/SaveBookDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/BookDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
  /books/delete/{id}:
    delete:
      tags:
        - book
      summary: deleteBook
      operationId: deleteBookUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
...
  /books/updateBook:
    put:
      tags:
        - book
      summary: updateBook
      operationId: updateBookUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: bookDTO
          description: bookDTO
          required: true
          schema:
            $ref: '#/definitions/SaveBookDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/BookDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
  /genres/getByName/{name}:
    get:
      tags:
        - genre
      summary: findByName
      operationId: findByNameUsingGET_1
      produces:
        - '*/*'
      parameters:
        - name: name
          in: path
          description: name
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/GenreDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /genres/updateGenre:
    put:
      tags:
        - genre
      summary: updateBook
      operationId: updateBookUsingPUT_1
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: genreDTO
          description: genreDTO
          required: true
          schema:
            $ref: '#/definitions/GenreDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/GenreDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /orders/addOrder:
    post:
      tags:
        - order
      summary: addOrder
      operationId: addOrderUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: orderDTO
          description: orderDTO
          required: true
          schema:
            $ref: '#/definitions/SaveOrderDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/OrderDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /orders/admin/updateOrder:
    post:
      tags:
        - order
      summary: updateOrderByAdmin
      operationId: updateOrderByAdminUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: newOrder
          description: newOrder
          required: true
          schema:
            $ref: '#/definitions/UpdateOrderDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/OrderDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /orders/allOrders:
    get:
      tags:
        - order
      summary: getAllOrders
      operationId: getAllOrdersUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/OrderDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /orders/delete/{id}:
    delete:
      tags:
        - order
      summary: deleteOrder
      operationId: deleteOrderUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
  /orders/getById/{id}:
    get:
      tags:
        - order
      summary: geOrderById
      operationId: geOrderByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/OrderDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /orders/updateOrder:
    post:
      tags:
        - order
      summary: updateOrder
      operationId: updateOrderUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: newOrder
          description: newOrder
          required: true
          schema:
            $ref: '#/definitions/UpdateOrderDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/OrderDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /publishers/addPublisher:
    post:
      tags:
        - publisher
      summary: addPublisher
      operationId: addPublisherUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: publisher
          description: publisher
          required: true
          schema:
            $ref: '#/definitions/PublisherDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PublisherDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /publishers/allPublishers:
    get:
      tags:
        - publisher
      summary: findAll
      operationId: findAllUsingGET_1
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/PublisherDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /publishers/deletePublisher/{id}:
    delete:
      tags:
        - publisher
      summary: deletePublisher
      operationId: deletePublisherUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
  /publishers/findByName/{name}:
    get:
      tags:
        - publisher
      summary: findByName
      operationId: findByNameUsingGET_2
      produces:
        - '*/*'
      parameters:
        - name: name
          in: path
          description: name
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/PublisherDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /publishers/getById/{id}:
    get:
      tags:
        - publisher
      summary: getPublisherById
      operationId: getPublisherByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PublisherDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /publishers/updatePublisher:
    put:
      tags:
        - publisher
      summary: updateBook
      operationId: updateBookUsingPUT_2
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: publisherDTO
          description: publisherDTO
          required: true
          schema:
            $ref: '#/definitions/PublisherDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PublisherDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /users/addUser:
    post:
      tags:
        - user
      summary: addBook
      operationId: addBookUsingPOST_1
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: userDTO
          description: userDTO
          required: true
          schema:
            $ref: '#/definitions/UserDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/UserDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /users/allUsers:
    get:
      tags:
        - user
      summary: getAllUsers
      operationId: getAllUsersUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/UserDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /users/delete/{id}:
    delete:
      tags:
        - user
      summary: deleteBook
      operationId: deleteBookUsingDELETE_1
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
  /users/getById/{id}:
    get:
      tags:
        - user
      summary: geUserById
      operationId: geUserByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/UserDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /users/updateUser:
    put:
      tags:
        - user
      summary: updateUser
      operationId: updateUserUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: userDTO
          description: userDTO
          required: true
          schema:
            $ref: '#/definitions/UserDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/UserDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
definitions:
  AuthorDTO:
    type: object
    properties:
      books:
        type: array
        items:
          $ref: '#/definitions/BookNameDTO'
      dateOfBirth:
        type: string
        format: date
      fullName:
        type: string
      genreList:
        type: array
        items:
          type: string
      id:
        type: integer
        format: int64
    title: AuthorDTO
...
  AuthorNameDTO:
    type: object
    properties:
      fullName:
        type: string
      id:
        type: integer
        format: int64
    title: AuthorNameDTO
  BookDTO:
    type: object
    properties:
      authors:
        type: array
        items:
          $ref: '#/definitions/AuthorNameDTO'
      bookId:
        type: integer
        format: int64
      deleted:
        type: boolean
      genres:
        type: array
        items:
          $ref: '#/definitions/GenreNameDTO'
      pageCount:
        type: integer
        format: int32
      price:
        type: integer
        format: int32
      publisher:
        type: string
      releaseYear:
        type: string
        format: date
      title:
        type: string
    title: BookDTO
...
  BookNameDTO:
    type: object
    properties:
      bookId:
        type: integer
        format: int64
      title:
        type: string
    title: BookNameDTO
  GenreDTO:
    type: object
    properties:
      genreName:
        type: string
      id:
        type: integer
        format: int64
    title: GenreDTO
  GenreNameDTO:
    type: object
    properties:
      genreName:
        type: string
    title: GenreNameDTO
...
  OrderDTO:
    type: object
    properties:
      books:
        type: array
        items:
          $ref: '#/definitions/BookNameDTO'
      order_id:
        type: integer
        format: int64
      status:
        type: string
      sum:
        type: integer
        format: int32
      user:
        type: string
    title: OrderDTO
  PublisherDTO:
    type: object
    properties:
      name:
        type: string
      publisherId:
        type: integer
        format: int64
    title: PublisherDTO
...
  SaveBookDTO:
    type: object
    properties:
      authorList:
        type: array
        items:
          type: integer
          format: int64
      bookId:
        type: integer
        format: int64
      deleted:
        type: boolean
      genreList:
        type: array
        items:
          type: integer
          format: int64
      pageCount:
        type: integer
        format: int32
      price:
        type: integer
        format: int32
      publisherId:
        type: integer
        format: int64
      releaseYear:
        type: string
        format: date
      title:
        type: string
    title: SaveBookDTO
  SaveOrderDTO:
    type: object
    properties:
      bookList:
        type: array
        items:
          type: integer
          format: int64
      orderId:
        type: integer
        format: int64
      status:
        type: string
      userId:
        type: array
        items:
          type: integer
          format: int64
    title: SaveOrderDTO
  UpdateOrderDTO:
    type: object
    properties:
      bookList:
        type: array
        items:
          type: integer
          format: int64
      orderId:
        type: integer
        format: int64
      status:
        type: string
      userId:
        type: integer
        format: int64
    title: UpdateOrderDTO
  UserDTO:
    type: object
    properties:
      blocked:
        type: boolean
      id:
        type: integer
        format: int64
      login:
        type: string
      password:
        type: string
      role:
        type: string
        enum:
          - USER
          - ADMIN
    title: UserDTO
...